requirements.txt
fastapi[all]==0.115
fastapi-users[sqlalchemy]
uvicorn[standard]~=0.34
alembic~=1.13.2
asyncpg~=0.29.0
sqlalchemy~=2.0.38
#psycopg2-binary
aiofiles
aiohttp
aioredis
jinja2
beautifulsoup4
plotly
rich~=13.7.1
pytest~=8.3.1
pytest_asyncio
slowapi~=0.1.9
sqladmin[full]==0.20.1
requests~=2.32
apscheduler~=3.10
fastapi-cache2[redis]
orjson

Faker~=27.0.0

# jwt
python-jose[cryptography]
pyjwt~=2.8.0
orjson~=3.10
starlette~=0.40
pydantic~=2.8.2
colorama~=0.4.6
python-dotenv~=1.0.1
httpx~=0.27.0
bcrypt~=4.1.2
pandas~=2.2.2
openpyxl~=3.1.5
gspread~=6.1.2

TgCrypto-pyrofork==1.2.7
pyrogram~=2.0.106